Added coefficient for video alignment in list
PinnedVideo  - added function which returns active pin offer for video

VideoTag - added code block which calculates coefficiant by views and
days which are left till the offer ends. Days are calucated using
getTimestamp which represents seconds in unix timestamp and then divided
by 60 to get minutes which are divided by 60 to get hours which are
divided by 24 to get days. At the end, ceil value of decimal day value
is calculated. If any data is missing coefficient is 0.